README.md
This repository is for the Student Rent It project. The project is hosted on pythonanywhere.com and has three branches - dev, test, prod. Changes can be done locally and then pushed to the remote dev branch. After those changes have been validated, the changes can be moved to the remote test branch and then pulled to the testenv branch on pythonanywhere for further testing. From there the changes can be pushed to the prod remote branch and then pulled into the prod branch on pythonanywhere.